refactor: extract credential secret logic from route into service
Move deriveCredentialSecret, getOrDeriveCredentialSecret, getCredentialSecret,
and resetCredentialSecret from the github-app route file into a dedicated
credential-secret-service. This fixes the inverted dependency where the task
worker was importing from a route file.
